edge-vision/requirements.txt
# AI Traffic Management System - Dependencies
# ==========================================

# Core AI & Machine Learning
stable-baselines3>=2.0.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
gymnasium>=0.28.0

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.7.0
Pillow>=9.0.0

# Traffic Simulation
sumo-rl>=1.4.0

# Data Science & Visualization
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Interface
Flask>=2.2.0
Flask-SocketIO>=5.3.0
python-socketio>=5.7.0

# Utilities
requests>=2.28.0
python-dateutil>=2.8.0
tqdm>=4.64.0

# Development & Testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# System Requirements
# ===================
# Python >= 3.8
# SUMO >= 1.12.0 (install separately)
# 
# Installation Instructions:
# -------------------------
# 1. pip install -r requirements.txt
# 2. Install SUMO from https://sumo.dlr.de/docs/Installing/
# 3. Download YOLOv8 models (automatic on first run)
#
# GPU Support (optional):
# ----------------------
# For CUDA support, install PyTorch with CUDA:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118